+++ src/types/assets.d.ts
@@ -7,3 +7,13 @@ declare module '*.yaml' {
   const content: string;
   export default content;
 }
+
+declare module '*.html' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.mjs' {
+  const content: string;
+  export default content;
+}